Use size_t when converting integer from/to pointer
Perviously we use uint32_t for this, but this doesn't compile for 64-bit
environment (and likely doesn't for 16-bit either.) Use size_t so that
we don't get size mismatch errors.

BUG=chrome-os-partner:19257
TEST=Run host emulated tests
BRANCH=None

+++ chip/host/system.c
@@ -63,9 +63,9 @@ int system_set_vbnvcontext(const uint8_t *block)
 	return EC_ERROR_UNIMPLEMENTED;
 }
 
-int system_usable_ram_end(void)
+uintptr_t system_usable_ram_end(void)
 {
-	return (int)(__shared_mem_buf + SHARED_MEM_SIZE);
+	return (uintptr_t)(__shared_mem_buf + SHARED_MEM_SIZE);
 }
 
 void system_pre_init(void)
